Updated generic parts symbol references
Added a hyphen to the front of the library reference for generic
components like resistors, caps. As I start adding more specific parts,
they will have their part number as the library symbol reference. Adding
the hyphen will keep the generics at the top of the list and make them
easy to find.