Added connectors, rotary encoder, and modified IC symbols
I made all IC symbols use the Comment parameter as the simplified part
number which is visible on the schematic. The BOM will have the full
part number as the the PartNumber. For example, a part may have extra
letters on the end to denote temperature range etc. In the case where a
part number is specific to a package, the PartNumber parameter and the
Comment should be the same full-length part number.