Updated Passives
Updated capacitor library to have most common caps and package sizes.
Modified the symbol names to make them easier to identify and so that
they sort from smallest to largest. Before, they were just part numbers
so it was hard to find a part sometimes. Most of the time I do not need
a specific cap on the BOM, but I still want to choose a capacitor
package that is common to reduce cost.